Add files via upload
brutebogo is a sorting algorithm that creates random lists of random sizes within a certain allowance, then checks to see if they are sorted. If they are sorted, the length of the list is compared to the random list. The sorting completes if this last comparison is a match. The assumption is made that the list in need of sorting is made of all integers between zero and the list length -1.

brutebogo check is a script which compares bogosort to brute bogo. Currently brutebogo is slightly more efficient that bogo with an efficiency value of O(n!). I believe I'm on the right track here, it may just be an logic implementation issue.